requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.3.4
django-coverage-plugin==1.5.0

Sphinx==1.5.5
django-extensions==1.7.8
Werkzeug==0.12.1
django-test-plus==1.0.17
factory-boy==2.8.1

django-debug-toolbar==1.7

# improved REPL
#ipdb==0.10.2

#pytest-django==3.1.2
#pytest-sugar==0.8.0
